airflow/requirements.txt
setuptools
sec-edgar-downloader
beautifulsoup4
html2text
lxml
boto3
python-dotenv
botocore

# For Pinecone and BioBERT implementation
pinecone-client
torch
transformers
numpy
tqdm
scikit-learn
requests
pandas